redefined all the lostpet functions to be xLostPetx instead of being the exact copy of spottedstray(previous animal. )
from the functioning navbar, redefined all the lostpet logic
