+++ client/app/components/Miscellaneous/ReportLost.js
@@ -136,7 +136,7 @@ const ReportLost = () => {
 
         try {
             const response = await fetch(
-                `${process.env.NEXT_PUBLIC_SERVER_URL}/api/lostpet`,
+                `${process.env.NEXT_PUBLIC_SERVER_URL}/api/spottedstray`,
                 {
                     method: 'POST',
                     headers: {
