+++ client/app/components/cards/FeaturedStrays.js
@@ -10,7 +10,7 @@ const FeaturedStrays = () => {
         const fetchAnimals = async () => {
             try {
                 const response = await fetch(
-                    `${process.env.NEXT_PUBLIC_SERVER_URL}/api/lostpet`
+                    `${process.env.NEXT_PUBLIC_SERVER_URL}/api/spottedstray`
                 )
                 const data = await response.json()
